Refactoring: cleff -> rio Introduced versioned config format (closes #21) Added tests for config parsing
