+++ .gitignore
@@ -32,3 +32,6 @@ tmp
 controller.tar.gz
 
 cli/licenses.md
+
+tags
+v011.json
