--- cli/Codchi/Platform/Windows.hs
@@ -1,876 +1,0 @@
-{-# LANGUAGE CPP #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE TypeOperators #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
-
-module Codchi.Platform.Windows (runCodchiLIO) where
-
-#ifdef mingw32_HOST_OS
-import Graphics.Win32 hiding (try)
-import System.Win32.Console.CtrlHandler
-import System.Win32.DLL (getModuleHandle)
-import System.Win32.Registry
-import System.Win32.Shell
-import System.Win32.Shortcut
-#else
-import Foreign (ForeignPtr, Ptr, nullPtr)
-import Foreign.C (CInt, CIntPtr, CWchar)
-#endif
-
-import Codchi.Dsl
-import Codchi.Types
-
-import Cleff hiding (send)
-import Cleff.Error
-import Codchi.Effects
-import Codchi.Parser
-import Control.Concurrent (threadDelay)
-import Control.Concurrent.Async
-import Control.Concurrent.STM (dupTChan, newTChan, tryReadTChan, writeTChan)
-import Control.Exception (try)
-import Control.Monad.Logger (LogLevel (..), logWithoutLoc)
-import qualified Data.Attoparsec.Text as P
-import qualified Data.ByteString as BS
-import Data.ByteString.Builder.Extra (defaultChunkSize)
-import qualified Data.ByteString.Lazy as BL
-import Data.Char (isSpace)
-import Data.List (stripPrefix)
-import qualified Data.String as String
-import qualified Data.Text as Text
-import qualified Data.Text.Lazy as LText
-import System.IO.Error (isPermissionError)
-import qualified System.IO.Utf8 as Utf8
-import System.Process.Typed (ExitCode (..), ProcessConfig)
-import qualified System.Process.Typed as Proc
-import UnliftIO (bracket, catch, finally, hClose, throwIO)
-import UnliftIO.Directory
-
-data WSLInstance = WSLInstance
-    { _wslName :: !Text
-    , _wslStatus :: !InstanceStatus
-    , _isWSL2 :: !Bool
-    }
-    deriving (Show)
-
-instance Parseable [WSLInstance] where
-    -- drop table header
-    parser = drop 1 <$> p
-      where
-        p = P.sepBy' lineP P.endOfLine <* P.endOfLine
-        lineP =
-            WSLInstance
-                <$> nameP
-                <*> isRunningP
-                <*> versionP
-                <* P.skipWhile (not . P.isEndOfLine)
-        nameP =
-            P.skipWhile (\x -> isSpace x || x == '*')
-                *> P.takeWhile1 (not . isSpace)
-        isRunningP =
-            P.skipWhile isSpace
-                *> ( (\s -> if s == "Running" then Running else Stopped)
-                        <$> P.takeWhile1 (not . isSpace)
-                   )
-        versionP =
-            P.skipWhile isSpace
-                *> ((== "2") <$> P.takeWhile1 (not . isSpace))
-
-data Shell = Windows | InWSL deriving (Eq, Show)
-
--- \| exit failure
-newtype ShellException
-    = ShellException Text
-    deriving (Show)
-    deriving anyclass (Exception)
-
--- | Somethings wrong with controller
-newtype ControllerException
-    = ControllerException String
-    deriving (Show)
-    deriving anyclass (Exception)
-
-runCodchiLIO :: Interpreter CodchiL (Logger : IOE : Errors [DriverException, UserError, Panic])
-runCodchiLIO = interpret $ \case
-    GetStatus ->
-        getControllerStatus >>= \case
-            WSLNotInstalledOrNeedsUpdate ->
-                throwError $
-                    DriverException $
-                        ControllerException $
-                            "WSL is not installed or needs an update.\r\n"
-                                <> "Please run 'wsl --install --web-download --no-distribution' or "
-                                <> "download the latest version from 'https://github.com/microsoft/WSL/releases'.\r\n"
-                                <> "This will require admin privileges and also restarting your computer."
-            ControllerNotInstalled -> return CodchiNotInstalled
-            ControllerInstalled status -> return $ CodchiInstalled status
-    ControllerInit ->
-        liftIO (flip findFile "codchi\\controller.tar.gz" =<< getXdgDirectoryList XdgDataDirs) >>= \case
-            Nothing ->
-                throwError $ Panic $ UnrecoverableError "Couldn't find root filesystem for Codchi controller"
-            Just rootfs -> do
-                ctrlDir <- getOrCreateXdgDir XdgState (mkNTPath (_CONTROLLER_DIR :: Text))
-                let ctrlFile =
-                        logTraceId "controller dir" $
-                            toNTPath $
-                                ctrlDir </> fromList ["controller.tar.gz"]
-                copyFile rootfs (toString ctrlFile)
-
-                let wslImport =
-                        void $
-                            rethrowAll $
-                                readProcess_ $
-                                    wsl'exeCmd
-                                        [ "--import"
-                                        , _CONTROLLER
-                                        , toNTPath ctrlDir
-                                        , ctrlFile
-                                        ]
-                    cleanup = removeFile $ toString ctrlFile
-
-                wslImport `finally` cleanup
-    ControllerStart -> runCodchiLIO $ rethrowAll mainLoop
-    GetInstanceInfo name -> do
-        wslInstances <- listWSLInstances
-        let instances =
-                [ CodchiInstance instanceName _wslStatus
-                | WSLInstance{..} <- wslInstances
-                , instanceName <-
-                    maybeToList $ rightToMaybe . parse =<< Text.stripPrefix _INSTANCE_PREFIX _wslName
-                ]
-        case find ((== name) . (.name)) instances of
-            Nothing -> return (CodchiInstance{name, status = NotInstalled})
-            Just i -> return i
-    RunCtrlNixCmd printLog cmd -> rethrowAll $ do
-        let printWSL Win _ = pass
-            printWSL WSL txt = printOutIO txt
-            rp
-                | printLog = readProcessWith printWSL printWSL
-                | otherwise = readProcess
-        result <- rp $ controllerCmd cmd
-
-        case result of
-            Left err ->
-                case Text.splitOn "error: " err of
-                    [_, e] | not (Text.null e) -> return $ Left e
-                    _ -> throwError $ ShellException err
-            Right out -> return $ Right out
-    RunInstanceCmd printLog inst cmd -> rethrowAll $ do
-        let printWSL Win _ = pass
-            printWSL WSL txt = printOutIO txt
-            rp
-                | printLog = readProcessWith printWSL printWSL
-                | otherwise = readProcess
-        result <- rp $ wslCmd inst.withPrefix cmd
-
-        case result of
-            Left err -> return $ Left err
-            Right out -> return $ Right out
-    DriverInstallInstance name rootfsPath -> do
-        instanceDir <-
-            liftIO $
-                getOrCreateXdgDir XdgState $
-                    mkNTPath ("instances" :: Text) </> mkNTPath name.text
-
-        putTextLn $ "Installing " <> show name.text <> " from " <> toText rootfsPath
-
-        void $
-            rethrowAll $
-                readProcess_ $
-                    wsl'exeCmd
-                        [ "--import"
-                        , name.withPrefix
-                        , toNTPath instanceDir
-                        , toText rootfsPath
-                        ]
-    DriverUninstallInstance name _status -> do
-        putTextLn $ "Uninstalling " <> name.text
-
-        void $
-            rethrowAll
-                (readProcess_ $ wsl'exeCmd ["--unregister", name.withPrefix])
-
-        liftIO $ do
-            startMenu <- getFolderPath cSIDL_PROGRAMS
-            let instanceFolder = startMenu </> fromList [_APP_NAME <> " - " <> name.text]
-            retryOnPermissionErrors 100 $
-                removePathForcibly (toString $ toNTPath instanceFolder)
-            refreshIconCache
-    RunInInstance i showTerm args -> do
-        liftIO $ do
-            consoleWindow <- getConsoleWindow
-            case consoleWindow of
-                Just hwnd | not showTerm -> failIfFalse_ "hide console window" $ showWindow hwnd sW_HIDE
-                _ -> pass
-        let wrapIO
-                | showTerm =
-                    Proc.setStdin (Proc.useHandleOpen stdin)
-                        . Proc.setStdout (Proc.useHandleOpen stdout)
-                        . Proc.setStderr (Proc.useHandleOpen stderr)
-                | otherwise =
-                    Proc.setStdin Proc.nullStream
-                        . Proc.setStdout Proc.nullStream
-                        . Proc.setStderr Proc.nullStream
-
-        Proc.runProcess_ $
-            wrapIO $
-                wsl'exeCmd $
-                    [ "-d"
-                    , i.name.withPrefix
-                    , "--shell-type"
-                    , "login"
-                    ]
-                        <> args
-    UpdateShortcuts name swSharePath -> runCodchiLIO $ rethrowAll $ Codchi.Platform.Windows.updateShortcuts name swSharePath
-    GetDriverModule -> return "driver-wsl"
-    GetDriverPath dirType s ->
-        toString . toNTPath . (</> s)
-            <$> case dirType of
-                DirCtrl -> do
-                    let dir = getWSLInstanceDir _CONTROLLER
-                    createDirectoryIfMissing True $ toString $ toNTPath dir
-                    return dir
-                DirState -> getOrCreateXdgDir XdgState emptyPath
-                DirConfig -> getOrCreateXdgDir XdgConfig emptyPath
-    GetControllerPath path ->
-        case "C:\\" `stripPrefix` path of
-            Just p -> return (Right $ fromList ["mnt", "c"] </> mkNTPath p)
-            Nothing -> return (Left "Please provide an absolute path located on C:\\")
-
-mainLoop :: [IOE, Error ShellException, Error Panic, Logger, CodchiL] :>> es => Eff es ()
-mainLoop = do
-    pulseaudioLogFile <- getDriverPath DirState $ fromList ["pulseaudio.log"]
-    let pulseaudioLogLevel = case logLevel of
-            LevelError -> 0 :: Int
-            LevelWarn -> 1
-            LevelDebug -> 4
-            _ -> 3
-    vcxsrvLogFile <- getDriverPath DirState $ fromList ["vcxsrv.log"]
-    let vcxsrvLogLevel = case logLevel of
-            LevelError -> 0 :: Int
-            LevelWarn -> 1
-            LevelDebug -> 3
-            _ -> 2
-
-    cancelChan <- atomically newTChan
-
-    logInfo "Waking up WSL..."
-    void $ runProcessSilent $ controllerCmd "true"
-    wslVmId <-
-        fmap (logTraceId "wslVmId")
-            . maybe (throwError $ Panic $ UnrecoverableError "Can't retrieve WSL's VM ID...") return
-            =<< liftIO findWslVmId
-
-    let subprog name cmd args = do
-            myCancelChan <- atomically $ dupTChan cancelChan
-            void $
-                runProcessSilent $
-                    shellProc "taskkill.exe" ["/F", "/IM", name]
-            let doLog lvl _ txt = runTChanLoggingT $ logWithoutLoc name lvl txt
-            fix $ \loop -> do
-                runProcessWith (doLog LevelInfo) (doLog LevelWarn) (shellProc cmd args)
-                whenNothingM_ (atomically $ tryReadTChan myCancelChan) $ do
-                    logIO LevelWarn $ name <> " exited unexpectedly. Restarting..."
-                    threadDelay 1_000_000
-                    loop
-
-    void $ liftIO $ do
-        winFolder <- toNTPath <$> getFolderPath cSIDL_PROGRAM_FILESx86
-        async $
-            subprog
-                "codchi_pulseaudio.exe"
-                (winFolder <> "\\PulseAudio\\bin\\codchi_pulseaudio.exe")
-                [ "--log-target=file:" <> toText pulseaudioLogFile
-                , "--log-time"
-                , "--log-level=" <> show pulseaudioLogLevel
-                , "--disallow-exit"
-                , "--disallow-module-loading"
-                , -- , "--system"
-                  "--exit-idle-time=-1"
-                ]
-    void $ liftIO $ do
-        winFolder <- toNTPath <$> getFolderPath cSIDL_PROGRAM_FILES
-        async $
-            subprog
-                "codchi_vcxsrv.exe"
-                (winFolder <> "\\VcXsrv\\codchi_vcxsrv.exe")
-                [ "-ac" -- disable access control
-                , "-noreset" -- dont restart after last client exits
-                , "-wgl" -- native opengl
-                , "-compositewm" -- previews for windows
-                , "-notrayicon"
-                , "-dpi"
-                , "auto"
-                , "-multiwindow" -- seamless mode
-                , "-clipboard"
-                , "-noprimary"
-                , "-logfile"
-                , toText vcxsrvLogFile
-                , "-logverbose"
-                , show vcxsrvLogLevel
-                , "-vmid"
-                , "{" <> toText wslVmId <> "}"
-                , "-vsockport"
-                , "6000"
-                ]
-    -- https://sourceforge.net/p/vcxsrv/discussion/986201/thread/1ab552d067/
-
-    void $ liftIO $ async $ do
-        myCancelChan <- atomically $ dupTChan cancelChan
-        void $ readProcess $ controllerCmd "pkill nix-daemon || true"
-        let doLog lvl _ txt = runTChanLoggingT $ logWithoutLoc "controller" lvl txt
-        fix $ \loop -> do
-            runProcessWith (doLog LevelInfo) (doLog LevelWarn) $
-                Proc.setStdin Proc.closed $
-                    controllerCmd "/bin/ctrl-serve"
-            whenNothingM_ (atomically $ tryReadTChan myCancelChan) $ do
-                doLog LevelWarn Win ("Codchi controller exited unexpectedly (please see logs). Restarting..." :: Text)
-                loop
-
-    runWinLoop $ do
-        putTextLn "Stopping controller..."
-        atomically $ writeTChan cancelChan ()
-        void $ runProcessSilent $ controllerCmd "pkill nix-daemon || true"
-        -- void $ runProcessSilent $ shellProc "taskkill.exe" [ "/F", "/IM", "codchi_pulseaudio.exe" ]
-        void $ runProcessSilent $ shellProc "taskkill.exe" ["/F", "/IM", "codchi_vcxsrv.exe"]
-
-runProcessSilent :: MonadIO m => ProcessConfig stdin stdout stderr -> m ExitCode
-runProcessSilent =
-    Proc.runProcess
-        . Proc.setStdout Proc.nullStream
-        . Proc.setStderr Proc.nullStream
-
-updateShortcuts ::
-    [IOE, Error ShellException, Error Panic, Logger, CodchiL] :>> es =>
-    CodchiName ->
-    FilePath ->
-    Eff es ()
-updateShortcuts name swSharePath = do
-    icosFolder <-
-        toString . toNTPath
-            <$> getOrCreateXdgDir XdgState (fromList ["icos", name.text])
-
-    -- cleanup start menu entries
-    mapM_ (removeFile . (\f -> icosFolder <> "\\" <> f)) =<< listDirectory icosFolder
-
-    desktopEntries <-
-        let parseDesktopEntryAndIco app = do
-                de <-
-                    parse @(DesktopEntry ()) . decodeUtf8
-                        <$> readFileBS (swSharePath <> "\\wsl\\applications\\" <> toString app <> ".desktop")
-
-                case de of
-                    Left err ->
-                        throwError $
-                            Panic $
-                                UnrecoverableError $
-                                    "Could not parse desktop entry for " <> app <> ":\n" <> err
-                    Right desktopEntry -> do
-                        let iconInCtrl = swSharePath <> "\\wsl\\icos\\" <> toString app <> ".ico"
-                            iconInWin = icosFolder <> "\\" <> toString desktopEntry.name <> ".ico"
-                        icon <-
-                            doesFileExist iconInCtrl >>= \case
-                                True -> do
-                                    copyFile iconInCtrl iconInWin
-                                    return $ Just iconInWin
-                                False -> do
-                                    log LevelWarn $ "Could not find .ico for " <> app
-                                    return Nothing
-                        return $ desktopEntry{icon}
-         in mapM parseDesktopEntryAndIco
-                . mapMaybe (Text.stripSuffix ".desktop" . toText)
-                =<< listDirectory (swSharePath <> "\\wsl\\applications")
-
-    startMenuFolder <- liftIO $ do
-        startMenu <- getFolderPath cSIDL_PROGRAMS
-        let instanceFolder = startMenu </> fromList [_APP_NAME <> " - " <> name.text]
-        createDirectoryIfMissing True (toString $ toNTPath instanceFolder)
-        return $ toString $ toNTPath instanceFolder
-
-    -- cleanup start menu entries
-    mapM_ (removeFile . (\f -> startMenuFolder <> "\\" <> f)) =<< listDirectory startMenuFolder
-
-    -- create shortcuts
-    -- currentDir <- getCurrentDirectory
-    homeDir <- getHomeDirectory
-    either (throwError . Panic . UnrecoverableError . show) return
-        =<< liftIO
-            ( runExceptT $ do
-                ExceptT initialize
-                forM_ desktopEntries $ \desktopEntry -> do
-                    let lnk =
-                            Shortcut
-                                { targetPath = _APP_NAME <> ".exe"
-                                , arguments =
-                                    String.unwords $
-                                        ["run"]
-                                            <> ["--no-terminal" | not desktopEntry.isTerminal]
-                                            <> [ toString name.text
-                                               , "--"
-                                               , toString desktopEntry.exec
-                                               ]
-                                , workingDirectory = homeDir
-                                , showCmd = if desktopEntry.isTerminal then ShowNormal else ShowMinimized
-                                , description = ""
-                                , iconLocation = (fromMaybe "" desktopEntry.icon, 0)
-                                , hotkey = 0
-                                }
-                        lnkPath = startMenuFolder <> "\\" <> toString desktopEntry.name <> ".lnk"
-                    ExceptT $ writeShortcut lnk lnkPath
-                liftIO uninitialize
-            )
-    liftIO refreshIconCache
-
--- retryDuring :: MonadIO m => Integer -> Int -> m (Maybe e) -> m (Maybe e)
--- retryDuring timeFrame maxTries act
---     | maxTries < 1  = pure Nothing
---     | otherwise     = loop 1 . round =<< liftIO getPOSIXTime
---     where
---         loop n firstTryTime = do
---             res <- act
---             curTime <- round <$> liftIO getPOSIXTime
---             case res of
---                 Nothing -> return Nothing
---                 Just err
---                     | curTime - firstTryTime <= timeFrame -> if n < maxTries
---                                                                 then loop (n+1) firstTryTime
---                                                                 else return $ Just err
---                     | otherwise -> loop 0 curTime
-
--- https://tarma.com/support/im9/using/symbols/functions/csidls.htm
-cSIDL_PROGRAMS :: CSIDL
-cSIDL_PROGRAMS = 2
-cSIDL_PROGRAM_FILESx86 :: CSIDL
-cSIDL_PROGRAM_FILESx86 = 42
-
-getFolderPath :: CSIDL -> IO (Path Abs)
-getFolderPath csidl = mkNTPath <$> sHGetFolderPath nullPtr csidl nullPtr 0
-
-getOrCreateXdgDir :: MonadIO m => XdgDirectory -> Path Rel -> m (Path Abs)
-getOrCreateXdgDir xdg dir = liftIO $ do
-    path <- mkNTPath <$> getXdgDirectory xdg _APP_NAME
-    let innerPath = path </> dir
-    createDirectoryIfMissing True (logTraceId "create-dir" $ toString $ toNTPath innerPath)
-    return innerPath
-
-mkNTPath :: ToText txt => txt -> Path t
-mkNTPath = Path . filter (not . Text.null) . Text.splitOn "\\" . toText
-
-toNTPath :: Path t -> Text
-toNTPath (Path p) = Text.intercalate "\\" p
-
--- instance IsString (Path t) where
---     fromString = mkNTPath . toText
-
-rethrowAll ::
-    forall catchAllErr errs rest a.
-    ( catchAllErr ~ Error DriverException
-    , catchAllErr :> rest
-    , errs
-        ~ Errors
-            [ ParseException
-            , ShellException
-            -- , WinException
-            -- , WSLException
-            -- , DecodeException
-            ]
-    ) =>
-    Eff (errs ++ rest) a ->
-    Eff rest a
-rethrowAll = pluckEffects @catchAllErr @errs
-
-controllerCmd :: Text -> ProcessConfig () () ()
-controllerCmd = wslCmd _CONTROLLER
-
-wslCmd :: Text -> Text -> ProcessConfig () () ()
-wslCmd name cmd =
-    wsl'exeCmd $
-        [ "-d"
-        , name
-        , "--shell-type"
-        , "login"
-        , "--user"
-        , "root"
-        , -- , "--cd", "/root"
-          "bash"
-        , "-ic"
-        ]
-            <> (["-x" | logLevel <= LevelDebug])
-            <> [ show $ "printf " <> _MAGIC_UTF8_SEQ <> " | tee /dev/stderr ; " <> cmd
-               ]
-
-wsl'exeCmd :: [Text] -> ProcessConfig () () ()
-wsl'exeCmd = shellCmd "wsl.exe"
-
-shellCmd :: Text -> [Text] -> ProcessConfig () () ()
-shellCmd cmd args =
-    logTraceShowId "shellCmd" $
-        Proc.shell $
-            toString $
-                unwords $
-                    cmd : args
-
-shellProc :: Text -> [Text] -> ProcessConfig () () ()
-shellProc cmd args =
-    logTraceShowId "shellProc" $
-        Proc.proc (toString cmd) (map toString args)
-
-listWSLInstances :: [Error DriverException, IOE] :>> es => Eff es [WSLInstance]
-listWSLInstances = rethrowAll $ parseProcess_ @[WSLInstance] $ wsl'exeCmd ["-l", "-v"]
-
-data WinCodchiStatus
-    = WSLNotInstalledOrNeedsUpdate
-    | ControllerNotInstalled
-    | ControllerInstalled InstanceStatus
-
-runProcessWith ::
-    MonadUnliftIO m =>
-    (OutputType -> Text -> IO ()) ->
-    (OutputType -> Text -> IO ()) ->
-    ProcessConfig () () () ->
-    m ()
-runProcessWith outLogger errLogger proc = do
-    let proc' =
-            proc
-                & Proc.setStderr Proc.createPipe
-                & Proc.setStdout Proc.createPipe
-    Proc.withProcessTerm proc' $ \p -> do
-        let send logger h = loop Win
-              where
-                loop prevType = do
-                    bs <- BS.hGetSome h defaultChunkSize
-                    let txt = decodeUtf16And8 $ toLazy bs
-                        fixEOL
-                            | prevType == WSL = Text.replace "\n" "\r\n"
-                            | otherwise = id
-                    unless (BS.null bs) $
-                        case map fixEOL $ Text.splitOn _MAGIC_UTF8_SEQ txt of
-                            [win, wsl] -> do
-                                void $ logger Win win
-                                void $ logger WSL wsl
-                                loop WSL
-                            _ -> do
-                                void $ logger prevType txt
-                                loop prevType
-        void $ liftIO $ async $ send outLogger (Proc.getStdout p)
-        void $ liftIO $ async $ send errLogger (Proc.getStderr p)
-        void $ Proc.waitExitCode p
-
-data OutputType = Win | WSL deriving (Eq)
-
-readProcessWith ::
-    MonadUnliftIO m =>
-    (OutputType -> Text -> IO ()) ->
-    (OutputType -> Text -> IO ()) ->
-    ProcessConfig () () () ->
-    m (Either Text Text)
-readProcessWith outLogger errLogger proc = do
-    let proc' =
-            Proc.setStdout (loggingTextOutput outLogger) $
-                Proc.setStderr (loggingTextOutput errLogger) $
-                    logTraceShowId "readProcess" proc
-        collect p =
-            atomically $
-                (,,)
-                    <$> Proc.waitExitCodeSTM p
-                    <*> Proc.getStdout p
-                    <*> Proc.getStderr p
-        splitWinWSL txt = case Text.splitOn _MAGIC_UTF8_SEQ txt of
-            [_, wsl] | not (Text.null wsl) -> wsl
-            _ -> txt
-    (exitCode, out, err) <-
-        logTraceShowId "readProcess result"
-            <$> Proc.withProcessTerm proc' collect
-    case exitCode of
-        ExitSuccess -> return $ Right $ logTraceId "splitWinWSL" $ splitWinWSL $ toStrict out
-        ExitFailure _ -> return $ Left $ logTraceId "splitWinWSL" $ splitWinWSL $ toStrict err
-
-loggingTextOutput :: (OutputType -> Text -> IO ()) -> Proc.StreamSpec 'Proc.STOutput (STM LText)
-loggingTextOutput logger = Proc.mkPipeStreamSpec loggingTextOutputFromHandle
-  where
-    loggingTextOutputFromHandle :: ProcessConfig () () () -> Handle -> IO (STM LText, IO ())
-    loggingTextOutputFromHandle pc h = Utf8.withHandle h $ do
-        mvar <- newEmptyTMVarIO
-
-        void $ async $ do
-            let loop prevType front = do
-                    bs <- BS.hGetSome h defaultChunkSize
-                    let txt = decodeUtf16And8 $ toLazy bs
-                        fixEOL
-                            | prevType == WSL = Text.replace "\n" "\r\n"
-                            | otherwise = id
-                    if BS.null bs
-                        then atomically $ putTMVar mvar $ Right $ LText.fromChunks $ front []
-                        else case map fixEOL $ Text.splitOn _MAGIC_UTF8_SEQ txt of
-                            [win, wsl] -> do
-                                logger Win win
-                                logger WSL wsl
-                                loop WSL $ front . (wsl :) . (win :)
-                            _ -> do
-                                logger prevType txt
-                                loop prevType $ front . (txt :)
-            loop Win id `catch` \e -> do
-                atomically $ void $ tryPutTMVar mvar $ Left $ Proc.ByteStringOutputException e pc
-                throwIO e
-
-        return (readTMVar mvar >>= either throwSTM return, hClose h)
-
--- catchIf
---     :: (MonadUnliftIO m, Exception e)
---     => (e -> Bool) -> m a -> (e -> m a) -> m a
--- catchIf f a b = catch a (\e -> if f e then b e else throwIO e)
-
-readProcess :: MonadUnliftIO m => ProcessConfig () () () -> m (Either Text Text)
-readProcess = readProcessWith (\_ _ -> pass) (\_ _ -> pass)
-
-decodeUtf16And8 :: BL.ByteString -> Text
-decodeUtf16And8 = decodeUtf8 . BL.filter (/= 0)
-
-readProcess_ :: [IOE, Error ShellException] :>> es => ProcessConfig () () () -> Eff es Text
-readProcess_ = either (throwError . ShellException) return <=< readProcess
-
-parseProcess :: (MonadUnliftIO m, Parseable a) => ProcessConfig () () () -> m (Either Text a)
-parseProcess = fmap (parse =<<) . readProcess
-
-parseProcess_ ::
-    forall a es.
-    ( [Error ParseException, IOE] :>> es
-    , Parseable a
-    ) =>
-    ProcessConfig () () () ->
-    Eff es a
-parseProcess_ = either (throwError . ParseException) return <=< parseProcess
-
-getControllerStatus :: [Error Panic, IOE] :>> es => Eff es WinCodchiStatus
-getControllerStatus =
-    readProcess "wsl.exe --version" >>= \case
-        -- WSL is not activated or needs an update (--version is only available
-        -- in later versions)
-        Left _ -> return WSLNotInstalledOrNeedsUpdate
-        Right _ ->
-            readProcess "wsl.exe -l -v" >>= \case
-                Right out -> do
-                    instances <-
-                        either (throwError . Panic . ParseException) return $
-                            parse @[WSLInstance] out
-                    case find ((== _CONTROLLER) . (._wslName)) instances of
-                        Nothing -> return ControllerNotInstalled
-                        Just status -> return $ ControllerInstalled status._wslStatus
-                -- WSL outputs an error message to stdout if no distribution is
-                -- installed... We assume that if there is another error, there
-                -- will be an error message
-                Left err | Text.null err -> return ControllerNotInstalled
-                Left err -> throwError $ Panic $ UnrecoverableError $ logTraceId "wsl -l -v" err
-
-getWSLInstanceDir :: Text -> Path Abs
-getWSLInstanceDir instanceName = fromList ["\\\\wsl$", instanceName]
-
-_MAGIC_UTF8_SEQ :: IsString s => s
-_MAGIC_UTF8_SEQ = "this_is_for_recognizing_utf8"
-
--- splitMagicUTF8Seq txt = case Text.splitOn _MAGIC_UTF8_SEQ txt of
---     [win, wsl] -> (win, wsl)
---     _ | _MAGIC_UTF8_SEQ `Text.isInfixOf` txt -> ("", txt)
-
--- https://otter-o.hatenadiary.org/entry/20090217/1234861028
-runWinLoop :: MonadIO m => IO () -> m ()
-runWinLoop cleanup = liftIO $ do
-    withConsoleCtrlHandler (\_ -> cleanup >> exitSuccess >> pure True) $ do
-        void $ createMessageWindow msgLoop
-        allocaMessage messagePump
-  where
-    createMessageWindow wndProc = do
-        let clsName = mkClassName _APP_NAME
-        hinst <- getModuleHandle Nothing
-        whenNothingM_ (registerClass (0, hinst, Nothing, Nothing, Nothing, Nothing, clsName)) $
-            error "Couldn't register window class"
-        createWindow
-            clsName
-            _APP_NAME
-            0
-            Nothing
-            Nothing
-            Nothing
-            Nothing
-            (Just nullPtr)
-            Nothing
-            hinst
-            wndProc
-    msgLoop hwnd msg wp lp
-        | msg == wM_CLOSE = cleanup >> exitSuccess >> return 0
-        | otherwise = liftIO $ defWindowProc (Just hwnd) msg wp lp
-    messagePump msg = void $ infinitely $ do
-        whenM (getMessage msg Nothing) $ do
-            void $ translateMessage msg
-            void $ dispatchMessage msg
-
-findWslVmId :: IO (Maybe String)
-findWslVmId = do
-    let vmsKey = regOpenKeyEx hKEY_LOCAL_MACHINE "SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion\\HostComputeService\\VolatileStore\\ComputeSystem" kEY_READ
-    withRegKey vmsKey $ \hkey -> do
-        vmKeys <- regEnumKeys hkey
-
-        let filterWslVms vmid = do
-                subkeys <- withRegKey (regOpenKeyEx hkey vmid kEY_READ) regEnumKeyVals
-
-                return $
-                    if any isWslVm subkeys
-                        then Just vmid
-                        else Nothing
-
-        asum <$> mapM filterWslVms vmKeys
-  where
-    isWslVm ("ComputeSystemType", "2", _) = True
-    isWslVm _ = False
-
-withRegKey :: IO HKEY -> (HKEY -> IO c) -> IO c
-withRegKey aquire = bracket aquire regCloseKey
-
-retryOnPermissionErrors :: Int -> IO a -> IO a
-retryOnPermissionErrors maxRetries io = do
-    res <- try io
-    case res of
-        Right a -> return a
-        Left ex
-            | isPermissionError ex && maxRetries > 1 -> do
-                let retries' = logTrace "retryOnPermissionErrors" (show @Text (ex, maxRetries)) (maxRetries - 1)
-                threadDelay ((maxRetries - retries') * 200)
-                retryOnPermissionErrors retries' io
-            | otherwise -> throwIO $ logTraceShowId "retryOnPermissionErrors error" ex
-
-#ifdef mingw32_HOST_OS
-foreign import ccall "shlobj_core.h SHChangeNotify"
-    c_SHChangeNotify :: LONG -> UINT32 -> LPVOID -> LPVOID -> IO ()
-
-refreshIconCache :: IO ()
-refreshIconCache = c_SHChangeNotify 0x8000000 0x1000 nullPtr nullPtr
-
--- foreign import ccall "windows.h FreeConsole"
---     c_FreeConsole :: IO BOOL
-
--- freeConsole :: IO ()
--- freeConsole = failIfFalse_ "FreeConsole" c_FreeConsole
-
-foreign import ccall "windows.h GetConsoleWindow"
-    c_GetConsoleWindow :: IO HWND
-
-getConsoleWindow :: IO (Maybe HWND)
-getConsoleWindow = do
-    hwnd <- c_GetConsoleWindow
-    return $ if hwnd /= nullPtr
-        then Just hwnd
-        else Nothing
-
-#else
-
--- fake functions for HLS under linux
-
-noop :: a
-noop = error "Not implemented on linux"
-
-type DWORD = Word32
-type WindowStyle = DWORD
-type LONG = Int32
-type LPMSG = Ptr ()
-type HWND = Ptr ()
-type HMENU = Ptr ()
-type HANDLE = Ptr ()
-type HINSTANCE = Ptr ()
-type LPTSTR = Ptr CWchar
-type WindowClosure = HWND -> DWORD -> Word -> CIntPtr -> IO CIntPtr
-type REGSAM = Word32
-type HKEY = ForeignPtr ()
-type CSIDL = CInt
-
-wM_CLOSE :: DWORD
-wM_CLOSE = noop
--- wM_DESTROY :: DWORD
--- wM_DESTROY = noop
-
-failIfFalse_ :: String -> IO Bool -> IO ()
-failIfFalse_ = noop
-
-withConsoleCtrlHandler :: (DWORD -> IO Bool) -> IO a -> IO a
-withConsoleCtrlHandler = noop
-
-getMessage :: LPMSG -> Maybe HWND -> IO Bool
-getMessage = noop
-allocaMessage :: (LPMSG -> IO a) -> IO a
-allocaMessage = noop
-translateMessage :: LPMSG -> IO Bool
-translateMessage = noop
-dispatchMessage :: LPMSG -> IO LONG
-dispatchMessage = noop
-defWindowProc :: Maybe HWND -> DWORD -> Word -> CIntPtr -> IO CIntPtr
-defWindowProc = noop
--- postQuitMessage :: Int -> IO ()
--- postQuitMessage = noop
-
-mkClassName :: String -> a
-mkClassName = noop
-registerClass :: (Word32, HINSTANCE, Maybe HANDLE, Maybe HANDLE, Maybe HANDLE, Maybe LPTSTR, LPTSTR) -> IO (Maybe Word16)
-registerClass = noop
-createWindow
-  :: LPTSTR -> String -> WindowStyle ->
-     Maybe Int -> Maybe Int -> Maybe Int -> Maybe Int ->
-     Maybe HWND -> Maybe HMENU -> HINSTANCE -> WindowClosure ->
-     IO HWND
-createWindow = noop
-
-type ShowWindowControl = DWORD
-sW_HIDE :: ShowWindowControl
-sW_HIDE = noop
-showWindow :: HWND -> ShowWindowControl -> IO Bool
-showWindow = noop
-
-getModuleHandle :: Maybe String -> IO HINSTANCE
-getModuleHandle = noop
-
-hKEY_LOCAL_MACHINE :: HKEY
-hKEY_LOCAL_MACHINE = noop
-kEY_READ :: REGSAM
-kEY_READ = noop
-regOpenKeyEx :: HKEY -> String -> REGSAM -> IO HKEY
-regOpenKeyEx = noop
-regCloseKey :: HKEY -> IO ()
-regCloseKey = noop
-regEnumKeys :: HKEY -> IO [String]
-regEnumKeys = noop
-regEnumKeyVals :: HKEY -> IO [(String,String,DWORD)]
-regEnumKeyVals = noop
-
-cSIDL_PROGRAM_FILES :: CSIDL
-cSIDL_PROGRAM_FILES = 38
-type SHGetFolderPathFlags = DWORD
-sHGetFolderPath :: HWND -> CSIDL -> HANDLE -> SHGetFolderPathFlags -> IO String
-sHGetFolderPath = noop
-
--- win32-shortcut
-
-data ShowCmd = ShowNormal | ShowMaximized | ShowMinimized deriving (Show)
-data Shortcut = Shortcut
-    { targetPath       :: FilePath
-    , arguments        :: String
-    , workingDirectory :: FilePath
-    , showCmd          :: ShowCmd
-    , description      :: String
-    , iconLocation     :: (FilePath, Int)
-    , hotkey           :: DWORD
-    }
-  deriving (Show)
-
-type ShortcutError = Void
-initialize :: IO (Either ShortcutError ())
-initialize = noop
-uninitialize :: IO ()
-uninitialize = noop
-writeShortcut :: Shortcut -> FilePath -> IO (Either ShortcutError ())
-writeShortcut = noop
-refreshIconCache :: IO ()
-refreshIconCache = noop
-
--- freeConsole :: IO ()
--- freeConsole = noop
-getConsoleWindow :: IO (Maybe HWND)
-getConsoleWindow = noop
-#endif
