+++ cli/src/Codchi.hs
@@ -1,89 +1,79 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeOperators #-}
 
-#if defined(mingw32_HOST_OS)
-import Codchi.Platform.Windows (runCodchiLIO)
-#elif (darwin_HOST_OS)
-import Codchi.Platform.Darwin (runCodchiLIO)
-#else
-import Codchi.Platform.Linux (runCodchiLIO)
-#endif
+module Codchi where
 
-import Cleff
-import Cleff.Error
 import Codchi.CLI
 import Codchi.Config
-import Codchi.Dsl
-import Codchi.Effects
+import Codchi.Config.IO
+import Codchi.Error
 import Codchi.Nix
-import Codchi.Parser (parse)
+import Codchi.Parser (parse, parse_)
+import Codchi.Platform
 import Codchi.Types
-import Control.Concurrent.Async (concurrently, wait, withAsync)
-import Control.Concurrent.STM (TChan, dupTChan, readTChan, writeTChan)
-import Control.Monad.Logger (LogLine, MonadLogger, monadLoggerLog, runFileLoggingT, runStdoutLoggingT)
+import qualified Control.Exception.Annotated.UnliftIO as Ann
 import Data.List (maximum)
 import Data.Map.Strict ((!?))
 import qualified Data.Map.Strict as Map
 import qualified Data.Text as Text
-import Main.Utf8 (withUtf8)
-import System.Exit (ExitCode (..))
-import UnliftIO.Directory (copyFile, createDirectoryIfMissing, doesFileExist, getCurrentDirectory)
 import Development.GitRev (gitHash)
+import RIO hiding (atomically, show, unlessM, unlines, whenM)
+import UnliftIO.Directory (copyFile, createDirectoryIfMissing, doesFileExist, getCurrentDirectory)
 
 _GIT_COMMIT :: String
-_GIT_COMMIT = logTraceShowId "current git hash" $(gitHash)
+_GIT_COMMIT = $(gitHash)
 
-cli ::
-    [Logger, CodchiL, IOE] :>> es =>
-    Errors '[NixError, Panic] :>> es =>
-    Command ->
-    Eff es ()
+cli :: Command -> RIO Codchi ()
 cli CmdStart = do
     let doStart = do
-            printLnOut "Updating controller..."
+            logInfo "Updating controller..."
             -- we update to the git commit with which the codchi executable was compiled with
             -- to ensure that executable and controller are compatible
+            gitCommit <- logTraceId "current git hash" _GIT_COMMIT
             let cmd =
                     Text.intercalate
                         " && "
                         [ "cd /"
                         , "nix run github:aformatik/codchi/"
-                            <> toText _GIT_COMMIT
+                            <> toText gitCommit
                             <> "#controller-rootfs.passthru.createContents"
                         , "/bin/ctrl-update-profile"
                         ]
             _ <-
-                either (throwError . NixError) return . logTraceShowId "controller update"
-                    =<< runCtrlNixCmd True cmd
+                logTraceId "controller update"
+                    =<< runCtrlNixCmd_ StreamStd cmd
 
-            printLnOut "Starting controller..."
+            logInfo "Starting controller..."
             controllerStart
     getStatus >>= \case
-        CodchiInstalled _ -> doStart
         CodchiNotInstalled -> do
             controllerInit
             getStatus >>= \case
-                CodchiNotInstalled -> throwError (Panic $ UnrecoverableError "Couldn't install controller!")
-                CodchiInstalled _ -> doStart
+                CodchiNotInstalled -> Ann.throw (InternalPanic "Couldn't install controller!")
+                _ -> doStart
+        _ -> doStart
 cli CmdStatus =
     getStatus >>= \case
         CodchiNotInstalled ->
-            printErrExit (ExitFailure 1) $
+            logExit $
                 "The codchi controller is not installed. Please run '" <> _APP_NAME <> " start' to install it."
-        CodchiInstalled status -> do
-            printLnOut $ "Controller Status: " <> show status
+        status -> do
+            logInfo $
+                "Controller Status: " <> case status of
+                    CodchiStopped -> "Stopped"
+                    CodchiRunning -> "Running"
 
-            instances <- mapM (getInstanceInfo . (.name)) . toList . (.instances) =<< readConfig
+            instances <- listMachines
             if not (null instances)
                 then
-                    printLnOut $
-                        showTable
-                            "\t"
-                            ( ["NAME", "STATUS"]
-                                : [[i.name.text, show i.status] | i <- instances]
-                            )
-                else printLnOut "No code machines found!"
+                    logInfo $
+                        display $
+                            showTable
+                                "\t"
+                                ( ["NAME", "STATUS"]
+                                    : [[i.name.text, show i.status] | i <- instances]
+                                )
+                else logInfo "No code machines found!"
 cli (CmdAddCa name path) = do
     absolutePath <-
         if "." `isPrefixOf` path
@@ -98,31 +88,32 @@ cli (CmdAddCa name path) = do
     let certFile = certDir <> "/" <> toString name.text <> ".crt"
 
     whenM (doesFileExist certFile) $
-        printErrExit (ExitFailure 1) $
+        logExit $
             "Certificate " <> show name.text <> " already exists."
 
     copyFile absolutePath certFile
 
     _ <-
-        either (throwError . NixError) return . logTraceShowId "controller update certificates"
-            =<< runCtrlNixCmd True "/bin/ctrl-update-certs"
+        logTraceId "controller update certificates"
+            =<< runCtrlNixCmd_ StreamStd "/bin/ctrl-update-certs"
 
-    printLnOut $
-        "Added certificate " <> show name.text <> " from " <> toText absolutePath <> " to controller certifacte store."
+    logInfo $
+        "Added certificate " <> show name.text <> " from " <> fromString absolutePath <> " to controller certifacte store."
 cli (CmdInit instanceName) =
     modifyConfig $ \cfg ->
         case Map.lookup instanceName.text cfg.instances of
             Just _existing ->
-                printErrExit
-                    (ExitFailure 1)
-                    ("Code machine " <> instanceName.text <> " already exists.")
+                logExit $
+                    "Code machine " <> display instanceName.text <> " already exists."
             Nothing ->
                 return (cfg{instances = Map.insert instanceName.text (defaultInstance instanceName) cfg.instances})
 cli (CmdAddModule instanceName opts) = modifyInstance instanceName $ \i ->
     case i.modules !? opts.moduleName.text of
         Just m ->
-            printErrExit (ExitFailure 1) $
-                "Not overwriting existing module " <> m.name.text <> ". Remove manually with `codchi remove-module`"
+            logExit $
+                "Not overwriting existing module "
+                    <> display m.name.text
+                    <> ". Remove manually with `codchi remove-module`"
         Nothing -> do
             uri <- do
                 let remoteParsed = GitModule <$> parse opts.uri
@@ -135,11 +126,11 @@ cli (CmdAddModule instanceName opts) = modifyInstance instanceName $ \i ->
                             else localParsed <> remoteParsed
                 case parsed of
                     Left err ->
-                        printErrExit (ExitFailure 1) $
+                        logExit $
                             "Could not parse URI "
                                 <> show opts.uri
                                 <> ". Allowed formats: http(s), ssh, absolute file path. Specific error: "
-                                <> err
+                                <> fromString err
                     Right u -> return u
 
             let modul = Module{uri, moduleType = opts.moduleType, name = opts.moduleName, branchCommit = opts.branchCommit}
@@ -147,12 +138,16 @@ cli (CmdAddModule instanceName opts) = modifyInstance instanceName $ \i ->
             -- automatically try to follow codchis' nixpkgs if not set already
             nixpkgsFollows <-
                 if isJust i.nixpkgsFollows
-                    then return $ logTraceShowId "add-module: already set" i.nixpkgsFollows
+                    then logTraceId "add-module: already set" i.nixpkgsFollows
                     else do
                         follow <- getNixpkgsInput modul
                         case follow of
                             Just (ModuleNixpkgs m) ->
-                                printLnOut $ "Code machine " <> i.name.text <> " now follows nixpkgs from " <> m.text
+                                logInfo $
+                                    "Code machine "
+                                        <> display i.name.text
+                                        <> " now follows nixpkgs from "
+                                        <> display m.text
                             _ -> pass
                         return follow
 
@@ -165,14 +160,14 @@ cli (CmdAddModule instanceName opts) = modifyInstance instanceName $ \i ->
 cli (CmdRemoveModule instanceName moduleName) = modifyInstance instanceName $ \i ->
     case i.modules !? moduleName.text of
         Nothing ->
-            printErrExit (ExitFailure 1) $
-                "Code machine " <> i.name.text <> " has no module with name " <> show moduleName.text <> "."
+            logExit $
+                "Code machine " <> display i.name.text <> " has no module with name " <> show moduleName.text <> "."
         Just _ -> do
             nixpkgsFollows <-
                 if i.nixpkgsFollows == Just (ModuleNixpkgs moduleName)
                     then do
                         logWarn $
-                            "Code machine " <> show instanceName.text <> " followed " <> moduleName.text <> "s' nixpkgs. This will now default to codchis' nixpkgs."
+                            "Code machine " <> show instanceName.text <> " followed " <> display moduleName.text <> "s' nixpkgs. This will now default to codchis' nixpkgs."
                         return Nothing
                     else return i.nixpkgsFollows
             return
@@ -183,24 +178,34 @@ cli (CmdRemoveModule instanceName moduleName) = modifyInstance instanceName $ \i
 cli (CmdSetFollows instanceName follows) = modifyInstance instanceName $ \i ->
     case follows of
         CodchiNixpkgs -> do
-            printLnOut $ "Code machine " <> show instanceName.text <> " now follows codchis' nixpkgs."
+            logInfo $ "Code machine " <> show instanceName.text <> " now follows codchis' nixpkgs."
             return (i{nixpkgsFollows = Just follows})
         ModuleNixpkgs m ->
             case i.modules !? m.text of
                 Nothing ->
-                    printErrExit (ExitFailure 1) $
-                        "Code machine " <> instanceName.text <> " has no module with name " <> show m.text
+                    logExit $
+                        "Code machine "
+                            <> display instanceName.text
+                            <> " has no module with name "
+                            <> show m.text
                 Just modul -> do
                     nixpkgsFollows <- getNixpkgsInput modul
                     when (isJust nixpkgsFollows) $
-                        printLnOut ("Code machine " <> instanceName.text <> " now follows " <> m.text <> "s' nixpkgs.")
+                        logInfo $
+                            "Code machine "
+                                <> display instanceName.text
+                                <> " now follows "
+                                <> display m.text
+                                <> "s' nixpkgs."
                     return (i{nixpkgsFollows})
 cli (CmdRebuild instanceName) = do
     cfg <- readConfig
     case cfg.instances !? instanceName.text of
         Nothing ->
-            printErrExit (ExitFailure 1) $
-                "Code machine " <> instanceName.text <> " does not exist."
+            logExit $
+                "Code machine "
+                    <> display instanceName.text
+                    <> " does not exist."
         Just machine -> do
             instanceDir <- getDriverPath DirCtrl (fromList ["instances", instanceName.text])
             instanceFlake <- getDriverPath DirCtrl (fromList ["instances", instanceName.text, "flake.nix"])
@@ -210,7 +215,9 @@ cli (CmdRebuild instanceName) = do
                 Just u -> return u
                 Nothing -> do
                     logWarn $
-                        "Follow not set for " <> instanceName.text <> ". Defaulting to codchis' nixpkgs."
+                        "Follow not set for "
+                            <> display instanceName.text
+                            <> ". Defaulting to codchis' nixpkgs."
                     return CodchiNixpkgs
 
             createDirectoryIfMissing True instanceDir
@@ -219,94 +226,89 @@ cli (CmdRebuild instanceName) = do
             let cmd = "/bin/ctrl-install " <> instanceName.text
 
             tarballPath <-
-                either
-                    (throwError . NixError)
-                    (rethrowPanic @ParseException . parse_ @StorePath)
-                    =<< runCtrlNixCmd True cmd
+                parse_ @StorePath
+                    =<< runCtrlNixCmd_ StreamStd cmd
 
             rootfsFile <- getDriverPath DirCtrl (tarballPath.path </> fromList ["rootfs.tar"])
 
-            info <- getInstanceInfo instanceName
+            info <- findCodeMachine_ instanceName
 
             case info.status of
-                NotInstalled -> do
+                MachineNotInstalled -> do
                     unlessM (doesFileExist rootfsFile) $
-                        throwError (Panic $ NixError $ "Nix build didn't build " <> toText rootfsFile <> " as expected.")
-                    printLnOut "Registering with driver..."
+                        Ann.throw (InternalNixError $ "Nix build didn't build " <> rootfsFile <> " as expected.")
+                    logInfo "Registering with driver..."
                     driverInstallInstance instanceName rootfsFile
-                Running -> do
-                    printLnOut "Activating system..."
+                MachineRunning -> do
+                    logInfo "Activating system..."
                     let installCmd = "/nix/var/nix/profiles/system/bin/switch-to-configuration switch"
-                    void $
-                        either (throwError . NixError) return =<< runInstanceCmd True instanceName installCmd
-                Stopped -> pass
+                    runInstanceCmd StreamStd instanceName installCmd >>= \case
+                        Left err -> Ann.throw $ InternalNixError $ toString err
+                        _ -> pass
+                _ -> pass
 
-            printLnOut "Creating shortcuts..."
+            logInfo "Creating shortcuts..."
             updateShortcuts instanceName =<< getDriverPath DirCtrl =<< getInstanceSWSharePath instanceName
 
-            printLnOut $ "Successfully installed " <> instanceName.text
+            logInfo $ "Successfully installed " <> display instanceName.text
 cli (CmdRun instanceName showTerm args) = do
     unlessM isCtrlRunning $
-        printErrExit (ExitFailure 1) "Codchi controller is not running..."
+        logExit "Codchi controller is not running..."
 
     findInstance instanceName >>= \case
         Nothing ->
-            printErrExit (ExitFailure 1) $
-                "Code machine " <> instanceName.text <> " does not exist."
+            logExit $
+                "Code machine " <> display instanceName.text <> " does not exist."
         Just i -> runInInstance i showTerm args
 cli (CmdUninstall instanceName) = modifyConfig $ \c ->
     case c.instances !? instanceName.text of
         Nothing -> do
-            printErrExit (ExitFailure 1) $
-                "Code machine " <> instanceName.text <> " does not exist."
+            logExit $
+                "Code machine " <> display instanceName.text <> " does not exist."
         Just _ -> do
-            printLnOut $
-                "Do you really want to uninstall " <> instanceName.text <> "?. THIS WILL DELETE ALL OF YOUR DATA!"
-            printLnOut "Type 'yes' to confirm!"
+            logInfo $
+                "Do you really want to uninstall " <> display instanceName.text <> "?. THIS WILL DELETE ALL OF YOUR DATA!"
+            logInfo "Type 'yes' to confirm!"
             answer <- getLine
             if answer /= "yes"
-                then printErrExit (ExitFailure 1) "Aborted uninstall."
+                then logExit "Aborted uninstall."
                 else do
-                    instanceStatus <- getInstanceInfo instanceName <&> (.status)
-                    when (instanceStatus /= NotInstalled) $ 
+                    instanceStatus <- (.status) <$> findCodeMachine_ instanceName
+                    when (instanceStatus /= MachineNotInstalled) $
                         driverUninstallInstance instanceName instanceStatus
-                    void $
-                        either (throwError . NixError) return
-                            =<< runCtrlNixCmd False ("rm -r /instances/" <> instanceName.text <> " || true")
+                    _ <-
+                        runCtrlNixCmd_ StreamIgnore $
+                            "rm -r /instances/" <> instanceName.text <> " || true"
                     return c{instances = Map.delete instanceName.text c.instances}
 
-findInstance :: (CodchiL :> es, IOE :> es) => CodchiName -> Eff es (Maybe InstanceConfig)
+findInstance :: MonadCodchi m => CodchiName -> m (Maybe InstanceConfig)
 findInstance name = do
     cfg <- readConfig
     return (cfg.instances !? name.text)
 
-modifyInstance ::
-    [CodchiL, IOE, Logger] :>> es =>
-    CodchiName ->
-    (InstanceConfig -> Eff es InstanceConfig) ->
-    Eff es ()
+modifyInstance :: CodchiName -> (InstanceConfig -> RIO Codchi InstanceConfig) -> RIO Codchi ()
 modifyInstance name f = modifyConfig $ \cfg ->
     case Map.lookup name.text cfg.instances of
         Just i -> do
             i' <- f i
             return (cfg{instances = Map.insert name.text i' cfg.instances})
         Nothing ->
-            printErrExit (ExitFailure 1) $ "Code machine " <> name.text <> " does not exist."
+            logExit $ "Code machine " <> display name.text <> " does not exist."
 
-getNixpkgsInput :: (CodchiL :> es) => Module -> Eff es (Maybe NixpkgsFollows)
+getNixpkgsInput :: Module -> RIO Codchi (Maybe NixpkgsFollows)
 getNixpkgsInput m = do
     let cmd = "nix flake metadata --no-write-lock-file --json '" <> toFlakeUrl m <> "' | jq '.locks.nodes | has(\"nixpkgs\")'"
-    output <- logTraceShowId "getNixpkgsInput" . fmap Text.strip <$> runCtrlNixCmd False cmd
+    output <-
+        logTraceId "getNixpkgsInput"
+            . fmap Text.strip
+            =<< runCtrlNixCmd StreamIgnore cmd
     case output of
         Right "true" -> do
             return (Just $ ModuleNixpkgs m.name)
         _ -> return Nothing
 
-isCtrlRunning :: CodchiL :> es => Eff es Bool
-isCtrlRunning =
-    getStatus <&> \case
-        CodchiInstalled status -> status == Running
-        _ -> False
+isCtrlRunning :: MonadCodchi m => m Bool
+isCtrlRunning = (== CodchiRunning) <$> getStatus
 
 showTable :: Text -> [[Text]] -> Text
 showTable sep t = unlines (map mkLine t)
@@ -319,38 +321,8 @@ showTable sep t = unlines (map mkLine t)
         Text.intercalate sep
             . zipWith (`Text.justifyLeft` ' ') colWidths
 
-getInstanceSWSharePath ::
-    Errors [NixError, Panic] :>> es =>
-    CodchiL :> es =>
-    CodchiName ->
-    Eff es (Path Rel)
+getInstanceSWSharePath :: MonadCodchi m => CodchiName -> m (Path Rel)
 getInstanceSWSharePath name = do
     let cmd = "readlink -f \"/nix/var/nix/profiles/per-instance/" <> name.text <> "/system/sw/share\""
-    either
-        (throwError . NixError)
-        (rethrowPanic @ParseException . fmap (.path) . parse_ @StorePath)
-        =<< runCtrlNixCmd False cmd
-
-main :: IO ()
-main = withUtf8 $ do
-    logFile <- runCodchiL $ getDriverPath DirState (fromList [_APP_NAME <> ".log"])
-
-    let stdoutSink = runStdoutLoggingT . unTChanLoggingT =<< atomically (dupTChan logChan)
-        fileSink = runFileLoggingT logFile . unTChanLoggingT =<< atomically (dupTChan logChan)
-
-    withAsync (stdoutSink `concurrently` fileSink) $ \action -> do
-        runCodchiL . cli =<< parseCmd
-        atomically $ writeTChan logChan Nothing
-        void $ wait action
-  where
-    runCodchiL =
-        runIOE
-            . interpretLogger
-            . pluckEffects @IOE @(Errors '[Panic, DriverException, NixError, UserError])
-            . runCodchiLIO
-
-unTChanLoggingT :: (MonadLogger m, MonadIO m) => TChan (Maybe LogLine) -> m ()
-unTChanLoggingT chan = fix $ \loop ->
-    atomically (readTChan chan) >>= \case
-        Just (loc, src, lvl, msg) -> monadLoggerLog loc src lvl msg >> loop
-        Nothing -> pass
+    fmap (.path) . parse_ @StorePath
+        =<< runCtrlNixCmd_ StreamIgnore cmd
