+++ cli/src/Codchi/Config.hs
@@ -1,0 +1,7 @@
+module Codchi.Config (
+    module Codchi.Config.Common,
+    module Codchi.Config.V012,
+) where
+
+import Codchi.Config.Common
+import Codchi.Config.V012
