+++ cli/test/Config/v012.json
@@ -1,0 +1,50 @@
+{
+  "test": {
+    "modules": {
+      "default": {
+        "branchCommit": null,
+        "moduleType": {
+          "contents": "configuration.nix",
+          "tag": "LegacyModule"
+        },
+        "name": "default",
+        "uri": {
+          "contents": "/mnt/c/Users/test/Desktop/nix",
+          "tag": "LocalModule"
+        }
+      }
+    },
+    "name": "test",
+    "nixpkgsFollows": null
+  },
+  "main": {
+    "modules": {
+      "default": {
+        "branchCommit": {
+          "branch": "main",
+          "commit": null
+        },
+        "moduleType": {
+          "contents": [
+            "test",
+            "development/codchi"
+          ],
+          "tag": "FlakeModule"
+        },
+        "name": "default",
+        "uri": {
+          "contents": {
+            "protocol": "GitHttps",
+            "uri": "oauth2:glpat-hef92ksl3Ab_j29sk2jR@gitlab.de/foo/bar"
+          },
+          "tag": "GitModule"
+        }
+      }
+    },
+    "name": "main",
+    "nixpkgsFollows": {
+      "contents": "default",
+      "tag": "CodchiNixpkgs"
+    }
+  }
+}
